🎨 🎨 [Refactor] App disbale font scaling, dark mode, minor refactor
