+++ src/screens/Details/details.tsx
