+++ src/screens/index.tsx
@@ -1,0 +1,4 @@
+export {Details} from './Details/details';
+export {Home} from './Home/home';
+export {Login} from './Login/login';
+export {Profile} from './Profile/profile';
